Tweet Bug Fixes

+++ src/controllers/tweet.controller.js
@@ -1,5 +1,5 @@
 import { ApiError } from "../utils/ApiError.js";
-import { ApiResponse } from "../utils/ApiResponse.js"
+import { ApiResponse } from "../utils/ApiResponse.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 import { User } from "../models/user.model.js";
 import { Tweet } from "../models/tweet.model.js";
@@ -17,26 +17,48 @@ const createTweet = asyncHandler(async (req, res) => {
     const newTweet = await Tweet.create({
         content: content,
         owner: userId,
-    }); 
+    });
 
     if (!newTweet) {
         throw new ApiError(500, "Internal Error in creating a new tweet");
     }
 
-    return res.status(201)
-        .json(new ApiResponse(201, {
-            tweet: newTweet,
-        }, "New Tweet Created"));
+    return res.status(201).json(
+        new ApiResponse(
+            201,
+            {
+                tweet: newTweet,
+            },
+            "New Tweet Created"
+        )
+    );
 });
 
 const updateTweet = asyncHandler(async (req, res) => {
     const { tweetId } = req.params;
     const { newContent } = req.body;
+    const user = req.user;
+
+    if (!tweetId || tweetId.trim() === "") {
+        throw new ApiError(400, "Tweet Id is required");
+    }
 
     if (!newContent || newContent.trim() === "") {
         throw new ApiError(400, "New Content is missing");
     }
-    
+
+    const tweet = await Tweet.findById(tweetId);
+
+    if (!tweet) {
+        throw new ApiError(400, "Wrong Tweet Id");
+    }
+
+    const owner = tweet.owner;
+
+    if (!user._id.equals(owner)) {
+        throw new ApiError(402, "User cant update other tweets");
+    }
+
     const updatedTweet = await Tweet.findByIdAndUpdate(
         tweetId,
         {
@@ -48,20 +70,41 @@ const updateTweet = asyncHandler(async (req, res) => {
             new: true,
         }
     );
-    
+
     if (!updatedTweet) {
-        throw new ApiError(403, "Tweet Id is Wrong");
+        throw new ApiError(500, "Internal Error in updating the tweet");
     }
 
-    return res.status(200)
-        .json(new ApiResponse(200, {
-            updatedTweet: updatedTweet,
-        }, "Tweet Updated"));
-    
+    return res.status(200).json(
+        new ApiResponse(
+            200,
+            {
+                updatedTweet: updatedTweet,
+            },
+            "Tweet Updated"
+        )
+    );
 });
 
 const deleteTweet = asyncHandler(async (req, res) => {
     const { tweetId } = req.params;
+    const user = req.user;
+
+    if (!tweetId || tweetId.trim() === "") {
+        throw new ApiError(400, "Tweet Id is required");
+    }
+
+    const tweet = await Tweet.findById(tweetId);
+
+    if (!tweet) {
+        throw new ApiError(400, "Tweet Id is wrong");
+    }
+
+    const owner = tweet.owner;
+
+    if (!user._id.equals(owner)) {
+        throw new ApiError(402, "User cant delete other tweets");
+    }
 
     const deletedTweet = await Tweet.findByIdAndDelete(tweetId);
 
@@ -86,9 +129,10 @@ const getTweets = asyncHandler(async (req, res) => {
     if (!username || username.trim() === "") {
         throw new ApiError(400, "Username Required");
     }
-    
+
     // aggregation pipeline
-    const userAggregate = await User.aggregate([ // returns an array
+    const userAggregate = await User.aggregate([
+        // returns an array
         {
             $match: {
                 username: username,
@@ -115,10 +159,39 @@ const getTweets = asyncHandler(async (req, res) => {
 
     const userTweets = userAggregate[0].userTweets; // this is again an array
 
-    return res.status(200)
-        .json(new ApiResponse(200, {
-            userTweets: userTweets,
-        }, "User Tweets Fetched Successfully"));
+    return res.status(200).json(
+        new ApiResponse(
+            200,
+            {
+                userTweets: userTweets,
+            },
+            "User Tweets Fetched Successfully"
+        )
+    );
 });
 
-export { createTweet, updateTweet, getTweets, deleteTweet };
+const getTweetById = asyncHandler(async (req, res) => {
+    const { tweetId } = req.params;
+
+    if (!tweetId || tweetId.trim() === "") {
+        throw new ApiError(400, "Tweet Id is required");
+    }
+
+    const tweet = await Tweet.findById(tweetId);
+
+    if (!tweet) {
+        throw new ApiError(400, "Wrong Tweet Id");
+    }
+
+    return res.status(200).json(
+        new ApiResponse(
+            200,
+            {
+                tweet: tweet,
+            },
+            "Tweet Fetched Successfully"
+        )
+    );
+});
+
+export { createTweet, updateTweet, getTweets, deleteTweet, getTweetById };
